Add refreshResults & Paginator (#53)
We rename a previously internal method _runInner to refreshResults which executes and loops until execution is complete (but does not fetch results). Additionally, we Introduce a new class Paginator which fetches and caches and index result sets by page. This class exposes an api similar to the Dune interface result page navigation:

nextPage, previousPage, lastPage and pageN.

Paginator can be constructed from the return value of refreshResults and a client of type ExecutionAPI.

Will want to add docs on how to use this, but the test introduced should give a pretty good idea of how to test it out in a front end.